refactor: 모든 api → app 폴더 안으로 이동
