+++ src/app/auth/auth.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { EmailModule } from 'src/email/email.module';
+import { EmailModule } from 'src/app/email/email.module';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
 import { UserEventsListener } from './event/user-events.listener';
