+++ src/app/auth/dto/verify-email.dto.ts
